+++ tests/testthat/test-attachment_link.R
@@ -1,7 +1,7 @@
 test_that("form_schema_parse warns not implemented", {
-    testthat::expect_warning(
-      form_schema_parse(list()),
-      "Not implemented."
+  testthat::expect_warning(
+    form_schema_parse(list()),
+    "Not implemented."
   )
 })
 
@@ -15,12 +15,13 @@ test_that("prefix_fn works", {
 
 test_that("attachment_link warns not implemented", {
   testthat::expect_warning(
-    attachment_link(tibble::tibble(a=1:10, b=20:30),
-                    Sys.getenv("ODKC_TEST_PID"),
-                    Sys.getenv("ODKC_TEST_FID"),
-                    url = Sys.getenv("ODKC_TEST_URL"),
-                    un = Sys.getenv("ODKC_TEST_UN"),
-                    pw = Sys.getenv("ODKC_TEST_PW")),
+    attachment_link(tibble::tibble(a = 1:10, b = 20:30),
+      Sys.getenv("ODKC_TEST_PID"),
+      Sys.getenv("ODKC_TEST_FID"),
+      url = Sys.getenv("ODKC_TEST_URL"),
+      un = Sys.getenv("ODKC_TEST_UN"),
+      pw = Sys.getenv("ODKC_TEST_PW")
+    ),
     "Not implemented."
   )
 })